backend/pyproject.toml
[project]
name = "donzo-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-ninja>=1.4.3",
]
